Add tests. Fix deletion of ols dirs.

+++ simpholib/Library.py
@@ -19,7 +19,7 @@ class Library(object):
             self.delete_old = True
         else:
             self.delete_old = False
-        self.folders_to_delete = []
+        self.folders_to_delete = set()
         
     def __str__(self):
         """Return a string containing a nicely printable representation of an object"""
@@ -57,7 +57,8 @@ class Library(object):
                     uid += 1
 
             if not dummy_dirname:
-                self.folders_to_delete.append(dummy_dirpath)
+                if not dummy_dirpath == self.src_path:
+                    self.folders_to_delete.add(dummy_dirpath)
 
         sys.stdout.flush()
         sys.stdout.write('\tDone!\n')
@@ -141,11 +142,15 @@ class Library(object):
             self.update_progress((number+1) / len_libraryset)
 
         if self.delete_old:
-            self.folders_to_delete.append(old_dir)
+            if not old_dir == self.src_path:
+                self.folders_to_delete.add(old_dir)
 
     def delete_old_folders(self):
-        for fol in self.folders_to_delete:
-            rmtree(fol)
+        if self.src_path in self.folders_to_delete:
+            raise Exception
+        else:
+            for fol in self.folders_to_delete:
+                rmtree(fol)
 
     @staticmethod
     def update_progress(progress):
